+++ tests/misc_tests/type_eq_ui_tests/stw_macro_privacy-err.rs
@@ -1,0 +1,14 @@
+use typewit::MakeTypeWitness;
+
+mod private {
+    typewit::simple_type_witness!{
+        enum Private {
+            U8 = u8
+        }
+    }    
+}
+
+
+fn main(){
+    let _: private::Private<u8> = MakeTypeWitness::MAKE;
+}