+++ tests/misc_tests/type_eq_ui_tests/stw_macro_type_trait_bounds-err.rs
@@ -9,8 +9,18 @@ typewit::simple_type_witness!{
     }
 }
 
+typewit::simple_type_witness!{
+    enum TypeParamsAngBrack<'a, T, U:, V: 'a, W: 'a + 'a, X: Debug, Y: 'a + Debug> {
+        U8 = (&'a (), T, U, V, W, X, Y) 
+    }
+}
+
 fn _foo<'a, T, U, V, W, X, Y>() {
     let _: TypeParams<'a, T, U, V, W, X, Y, _> = MakeTypeWitness::MAKE;
+}
+
+fn _bar<'a, T, U, V, W, X, Y>() {
+    let _: TypeParamsAngBrack<'a, T, U, V, W, X, Y, _> = MakeTypeWitness::MAKE;
 }
 
 fn main(){}